[TAS-584]fix: 설정 > 계정관리 메뉴 구현 및 기능 이관
