+++ src/screens/Mypage/Setting/Policy/index.tsx
@@ -5,9 +5,9 @@ import { BasicMenu } from 'components/Mypage/Setting/Menu';
 
 const Policy = () => (
   <View style={styles.container}>
-    <BasicMenu title="서비스 이용약관" />
-    <BasicMenu title="개인정보 처리방침" />
-    <BasicMenu title="마케팅(이벤트) 정보 수신 동의" />
+    <BasicMenu title="서비스 이용약관" isArrowVisible />
+    <BasicMenu title="개인정보 처리방침" isArrowVisible />
+    <BasicMenu title="마케팅(이벤트) 정보 수신 동의" isArrowVisible />
   </View>
 );
